refactor: [code-review] Delete the duplicated glob matcher: `orbit_comm… [ORB-11733]
[code-review] Delete the duplicated glob matcher: `orbit_common::fs::glob` is a verbatim copy of `orbit_types::policy::glob` with a split consumer base

Planned-By: codex
