--- crates/orbit-common/src/fs/glob.rs
@@ -1,185 +1,0 @@
-//! Glob-matching helpers shared by policy evaluation and artifact scopes.
-//!
-//! Originally lived as private helpers inside
-//! [`crate::types::policy_def`]; promoted here so `orbit-store`'s
-//! artifact stores can reuse the same matcher without taking a dependency
-//! on `orbit-policy` (which the architecture diagram forbids). The original
-//! call sites in `policy_def` now route through these public functions, so
-//! existing behavior is unchanged.
-//!
-//! # Semantics
-//! The grammar supports the three glob operators agents already see in
-//! `denyRead`/`denyModify` rules and `fsProfile` patterns:
-//! - `**` — any number of path segments (including zero).
-//! - `*` — any sequence of non-separator characters.
-//! - `?` — any single non-separator character.
-//!
-//! All inputs are normalized to forward-slash separators and stripped of
-//! leading `./`. Paths that escape the workspace (`..`, `~`, absolute) are
-//! rejected as `OrbitError::InvalidInput`.
-
-use std::path::{Component, Path};
-
-use regex::{Regex, RegexBuilder};
-
-use crate::OrbitError;
-
-/// Normalize a workspace-relative path for glob matching.
-///
-/// Trims whitespace, replaces backslashes with forward slashes, strips a
-/// leading `./`, and rejects paths that try to escape the workspace
-/// (`..`, `~`, absolute). Returns an empty string for the workspace root
-/// (`.`).
-///
-/// [ORB-10009] The result is rebuilt from the path's normal components, so
-/// redundant separators collapse to a canonical spelling: `a/./b`, `a//b`,
-/// and `a/b/` all normalize to `a/b`. Without this, a deny rule such as
-/// `secret/key.txt` could be dodged by spelling the same file
-/// `secret/./key.txt` while a broad allow rule (`**`) still matched.
-pub fn normalize_glob_path(path: &str) -> Result<String, OrbitError> {
-    let trimmed = path.trim();
-    if trimmed.is_empty() {
-        return Err(OrbitError::InvalidInput(
-            "filesystem path must not be empty".to_string(),
-        ));
-    }
-
-    let mut normalized = trimmed.replace('\\', "/");
-    while let Some(stripped) = normalized.strip_prefix("./") {
-        normalized = stripped.to_string();
-    }
-    if normalized == "." {
-        normalized.clear();
-    }
-    if normalized == "~"
-        || normalized.starts_with("~/")
-        || normalized.starts_with("../")
-        || normalized == ".."
-    {
-        return Err(OrbitError::InvalidInput(format!(
-            "filesystem path `{path}` must stay inside the workspace root"
-        )));
-    }
-
-    let path_ref = Path::new(&normalized);
-    if path_ref.is_absolute() {
-        return Err(OrbitError::InvalidInput(format!(
-            "filesystem path `{path}` must stay inside the workspace root"
-        )));
-    }
-
-    for component in path_ref.components() {
-        match component {
-            Component::CurDir | Component::Normal(_) => {}
-            Component::ParentDir | Component::RootDir | Component::Prefix(_) => {
-                return Err(OrbitError::InvalidInput(format!(
-                    "filesystem path `{path}` must stay inside the workspace root"
-                )));
-            }
-        }
-    }
-
-    // [ORB-10009] Canonical spelling: drop `.` segments, duplicate and
-    // trailing separators, so rule matching cannot be dodged by an
-    // equivalent-but-differently-spelled path.
-    Ok(join_normal_components(path_ref))
-}
-
-/// Rebuild a validated relative path from its `Normal` components, dropping
-/// `.` segments and redundant separators. The input must already be checked
-/// for `..` / root / prefix components. [ORB-10009]
-pub(crate) fn join_normal_components(path: &Path) -> String {
-    let mut rebuilt = String::new();
-    for component in path.components() {
-        if let Component::Normal(part) = component {
-            if !rebuilt.is_empty() {
-                rebuilt.push('/');
-            }
-            rebuilt.push_str(&part.to_string_lossy());
-        }
-    }
-    rebuilt
-}
-
-/// Test whether `rule` (a glob pattern) matches `path` (already normalized
-/// via [`normalize_glob_path`]).
-pub fn match_glob(rule: &str, path: &str) -> Result<bool, OrbitError> {
-    let regex = compile_glob_regex(rule).map_err(|error| {
-        OrbitError::InvalidInput(format!("invalid filesystem glob `{rule}`: {error}"))
-    })?;
-    Ok(regex.is_match(path))
-}
-
-/// Compile a glob pattern into an anchored regex.
-///
-/// Exposed as a building block for callers that want to cache compiled
-/// regexes across many path checks (e.g. evaluating a single rule against
-/// hundreds of candidate paths in a hot path).
-pub fn compile_glob_regex(rule: &str) -> Result<Regex, regex::Error> {
-    if rule == "." {
-        return compile_filesystem_regex(r"^$");
-    }
-
-    if let Some(prefix) = rule.strip_suffix("/**") {
-        if prefix.is_empty() {
-            return compile_filesystem_regex(r"^.*$");
-        }
-        // Route the prefix through the same segment-aware translation as the
-        // general path below. Escaping the whole prefix would turn any `*`/`?`
-        // in it into a literal, so `**/secrets/**` would compile to
-        // `^\*\*/secrets(?:/.*)?$` and match nothing — silently voiding a deny
-        // rule (H1). Translating instead yields `^(?:.*/)?secrets(?:/.*)?$`.
-        let body = translate_glob_body(prefix);
-        return compile_filesystem_regex(&format!("^{body}(?:/.*)?$"));
-    }
-
-    let body = translate_glob_body(rule);
-    compile_filesystem_regex(&format!("^{body}$"))
-}
-
-/// Translate a glob pattern into an *unanchored* regex body, honoring the
-/// segment-aware operators (`**/`, `**`, `*`, `?`) and escaping every other
-/// character. Callers wrap the result in anchors (`^`…`$`) and any suffix.
-fn translate_glob_body(rule: &str) -> String {
-    let chars: Vec<char> = rule.chars().collect();
-    let mut index = 0usize;
-    let mut pattern = String::new();
-    while index < chars.len() {
-        if chars[index] == '*' {
-            if index + 2 < chars.len() && chars[index + 1] == '*' && chars[index + 2] == '/' {
-                pattern.push_str("(?:.*/)?");
-                index += 3;
-                continue;
-            }
-            if index + 1 < chars.len() && chars[index + 1] == '*' {
-                pattern.push_str(".*");
-                index += 2;
-                continue;
-            }
-            pattern.push_str("[^/]*");
-            index += 1;
-            continue;
-        }
-
-        if chars[index] == '?' {
-            pattern.push_str("[^/]");
-            index += 1;
-            continue;
-        }
-
-        pattern.push_str(&regex::escape(&chars[index].to_string()));
-        index += 1;
-    }
-    pattern
-}
-
-fn compile_filesystem_regex(pattern: &str) -> Result<Regex, regex::Error> {
-    // L-0062: Match policy globs using the target filesystem's case identity.
-    RegexBuilder::new(pattern)
-        .case_insensitive(filesystem_globs_are_case_insensitive())
-        .build()
-}
-
-fn filesystem_globs_are_case_insensitive() -> bool {
-    cfg!(any(target_os = "macos", windows))
-}
