+++ crates/orbit-common/src/fs/tests/mod.rs
@@ -1,4 +1,3 @@
 mod git;
-mod glob;
 mod io;
 mod selector;
